Move field definition
